+++ Cargo.toml
@@ -122,7 +122,7 @@ petgraph = "0.6.3"
 pnet = "0.34"
 pnet_datalink = "0.34"
 proc-macro2 = "1.0.51"
-quinn = "0.10.1"
+quinn = {version = "0.10.1"}
 quote = "1.0.23"
 rand = { version = "0.8.5", default-features = false } # Default features are disabled due to usage in no_std crates
 rand_chacha = "0.3.1"
@@ -132,7 +132,7 @@ ron = "0.8.1"
 ringbuffer-spsc = "0.1.9"
 rsa = "0.9"
 rustc_version = "0.4.0"
-rustls = "0.22.2"
+rustls = {version = "0.23", features = ["ring"]}
 rustls-native-certs = "0.7.0"
 rustls-pemfile = "2.0.0"
 rustls-webpki = "0.102.0"
@@ -155,7 +155,7 @@ token-cell = { version = "1.4.2", default-features = false }
 tokio = { version = "1.35.1", default-features = false } # Default features are disabled due to some crates' requirements
 tokio-util = "0.7.10"
 tokio-tungstenite = "0.21"
-tokio-rustls = "0.25.0"
+tokio-rustls = "0.26.0"
 # tokio-vsock = see: io/zenoh-links/zenoh-link-vsock/Cargo.toml (workspaces does not support platform dependent dependencies)
 console-subscriber = "0.2"
 typenum = "1.16.0"
