+++ io/zenoh-link-commons/Cargo.toml
@@ -12,37 +12,49 @@
 #   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 #
 [package]
-rust-version = { workspace = true }
-name = "zenoh-link-commons"
-version = { workspace = true }
-repository = { workspace = true }
-homepage = { workspace = true }
 authors = { workspace = true }
-edition = { workspace = true }
-license = { workspace = true }
 categories = { workspace = true }
 description = "Internal crate for zenoh."
+edition = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+name = "zenoh-link-commons"
+repository = { workspace = true }
+rust-version = { workspace = true }
+version = { workspace = true }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
 compression = []
+tls = [
+  "base64",
+  "rustls",
+  "rustls-native-certs",
+  "rustls-pemfile",
+  "rustls-webpki",
+  "webpki-roots",
+]
 
 [dependencies]
 async-trait = { workspace = true }
-rustls = { workspace = true }
-rustls-webpki = { workspace = true }
+base64 = { workspace = true, optional = true }
 flume = { workspace = true }
 tracing = {workspace = true}
 serde = { workspace = true, features = ["default"] }
-secrecy = {workspace = true }
+tokio = { workspace = true, features = [
+  "fs",
+  "io-util",
+  "net",
+  "sync",
+  "time",
+] }
+tokio-util = { workspace = true, features = ["rt"] }
+webpki-roots = { workspace = true, optional = true }
 zenoh-buffers = { workspace = true }
 zenoh-codec = { workspace = true }
-zenoh-core = { workspace = true }
 zenoh-config = { workspace = true }
+zenoh-core = { workspace = true }
 zenoh-protocol = { workspace = true }
 zenoh-result = { workspace = true }
-zenoh-util = { workspace = true }
 zenoh-runtime = { workspace = true }
-tokio = { workspace = true, features = ["io-util", "net", "fs", "sync", "time"] }
-tokio-util = { workspace = true, features = ["rt"] }
-futures = { workspace = true }
+zenoh-util = { workspace = true }
