+++ io/zenoh-link-commons/src/lib.rs
@@ -21,7 +21,9 @@ extern crate alloc;
 
 mod listener;
 mod multicast;
+#[cfg(feature = "tls")]
 pub mod tls;
+
 mod unicast;
 
 use alloc::{borrow::ToOwned, boxed::Box, string::String, vec, vec::Vec};
