+++ io/zenoh-links/zenoh-link-quic/Cargo.toml
@@ -30,15 +30,11 @@ base64 = { workspace = true }
 futures = { workspace = true }
 tracing = {workspace = true}
 quinn = { workspace = true }
-rustls-native-certs = { workspace = true }
-rustls-pemfile = { workspace = true }
-rustls-webpki = { workspace = true }
-secrecy = {workspace = true }
 tokio = { workspace = true, features = ["io-util", "net", "fs", "sync", "time"] }
 tokio-util = { workspace = true, features = ["rt"] }
 zenoh-config = { workspace = true }
 zenoh-core = { workspace = true }
-zenoh-link-commons = { workspace = true }
+zenoh-link-commons = { workspace = true, features = ["tls"]}
 zenoh-protocol = { workspace = true }
 zenoh-result = { workspace = true }
 zenoh-sync = { workspace = true }
